Add new network monitor that listens and reacts to internet availability in real time (only checks the `isOnline` boolean and network interface property on the main thread - all monitoring activity itself is done off the main thread for optimal performance).
Add convenient app-wide access to internet availability via `appState.isOnline`. This allows us to bypass having to use the boilerplate`NetworkMonitor.shared` flow every time we want to check for online activity.

Local mode button now properly reacts to when internet access is available vs. not.

App now properly reacts with an "offline mode", barring entry into use of remote models when offline and providing user with options to load in Ollama through the auth flow.

Update auth flow with new network activity monitor.
