+++ macos/Onit/Constants/ModelConstants.swift
@@ -1,0 +1,10 @@
+//
+//  ModelConstants.swift
+//  Onit
+//
+//  Created by Loyd Kim on 7/7/25.
+//
+
+struct ModelConstants {
+    static let ollamaUrl: String = "http://localhost:11434"
+}
