+++ macos/Onit/UI/Auth/AuthFlow.swift
@@ -5,47 +5,19 @@
 //  Created by Loyd Kim on 4/30/25.
 //
 
-import AuthenticationServices
 import Defaults
-import GoogleSignIn
-import GoogleSignInSwift
 import SwiftUI
 
 struct AuthFlow: View {
     @Environment(\.appState) var appState
-    @Environment(\.openSettings) var openSettings
     
     @Default(.availableLocalModels) var availableLocalModels
     @Default(.authFlowStatus) var authFlowStatus
-    
-    @Default(.useOpenAI) var useOpenAI
-    @Default(.useAnthropic) var useAnthropic
-    @Default(.useXAI) var useXAI
-    @Default(.useGoogleAI) var useGoogleAI
-    @Default(.useDeepSeek) var useDeepSeek
-    @Default(.usePerplexity) var usePerplexity
-
-    @State private var isHoveredContinueWithEmailButton: Bool = false
-    @State private var isPressedContinueWithEmailButton: Bool = false
-    
-    @State private var isHoveredRedirectButton: Bool = false
-    @State private var isHoveredSkipButton: Bool = false
-
-    @State private var isHoveredResendLinkButton: Bool = false
-    @State private var isHoveredBackButton: Bool = false
-    @State private var isPressedBackButton: Bool = false
-    
-    @State private var isHoveredAddModelButton: Bool = false
     
     @State private var email: String = ""
     @State private var requestedEmailLogin: Bool = false
     
     @State private var errorMessageEmail: String = ""
-    @State private var errorMessageAuth: String = ""
-    
-    private var isSignUp: Bool {
-        return authFlowStatus == .showSignUp
-    }
     
     private var userProvidedOwnModel: Bool {
         let hasLocalModel: Bool = !availableLocalModels.isEmpty
@@ -59,14 +31,22 @@ struct AuthFlow: View {
     var body: some View {
         VStack(alignment: .center, spacing: 42) {
             if requestedEmailLogin {
-                emailAuthTokenForm
+                AuthVerifyEmailForm(
+                    email: $email,
+                    requestedEmailLogin: $requestedEmailLogin,
+                    requestEmailLoginLink: requestEmailLoginLink
+                )
             } else {
-                form
+                AuthForm(
+                    email: $email,
+                    errorMessageEmail: $errorMessageEmail,
+                    requestEmailLoginLink: requestEmailLoginLink
+                )
                 
-                redirectSection
+                AuthRedirectSection()
                 
                 if !userProvidedOwnModel {
-                    addModelSection
+                    AuthAddModelSection()
                 }
                 
                 Spacer()
@@ -82,194 +62,8 @@ struct AuthFlow: View {
         .frame(maxWidth: .infinity, maxHeight: .infinity)
         .padding(.top, 134)
     }
-}
-
-// MARK: - Child Components
-
-extension AuthFlow {
-    private var formAuthButtons: some View {
-        VStack(spacing: 4) {
-            OnboardingAuthButton(
-                icon: .logoGoogle,
-                action: handleGoogleSignInButton
-            )
-            
-            if !errorMessageAuth.isEmpty {
-                Text(errorMessageAuth)
-                    .styleText(size: 12, weight: .medium, color: .red, align: .center)
-            }
-        }
-    }
     
-    private var emailAddressInput: some View {
-        InputField(
-            placeholder: "Email Address",
-            text: $email,
-            errorMessage: errorMessageEmail,
-            onSubmit: requestEmailLoginLink
-        )
-    }
-    
-    private var continueWithEmailButton: some View {
-        Text("Continue with email")
-            .frame(maxWidth: .infinity)
-            .frame(height: 40)
-            .styleText(align: .center)
-            .addButtonEffects(
-                background: .blue400,
-                hoverBackground: .blue350,
-                cornerRadius: 9,
-                isHovered: $isHoveredContinueWithEmailButton,
-                isPressed: $isPressedContinueWithEmailButton,
-                action: requestEmailLoginLink
-            )
-    }
-    
-    private var form: some View {
-        VStack(alignment: .center, spacing: 0) {
-            Image("Logo").padding(.bottom, 19)
-            
-            Text(isSignUp ? "Create your account" : "Sign in to Onit")
-                .styleText(size: 23)
-                .padding(.bottom, 28)
-            
-            VStack(alignment: .center, spacing: 12) {
-                formAuthButtons
-                
-                Text("OR").styleText(size: 12, color: .gray300)
-                
-                VStack(spacing: 20) {
-                    VStack(spacing: 12) {
-                        emailAddressInput
-                        continueWithEmailButton
-                    }
-//                    agreement
-                }
-            }
-        }
-        .padding(.horizontal, 40)
-    }
-    
-    private var agreement: some View {
-        Text(generateAgreementText())
-            .styleText(size: 12, align: .center)
-    }
-    
-    private var redirectSection: some View {
-        HStack(spacing: 6) {
-            Text(isSignUp ? "Already have an account?" : "Don't have an account?")
-                .styleText(size: 13, weight: .regular, color: .gray100)
-            
-            Button {
-                if isSignUp {
-                    authFlowStatus = .showSignIn
-                } else {
-                    authFlowStatus = .showSignUp
-                }
-            } label: {
-                Text(isSignUp ? "Sign In" : "Sign up")
-                    .styleText(size: 13, weight: .regular, underline: isHoveredRedirectButton)
-            }
-            .buttonStyle(PlainButtonStyle())
-            .onHover { isHovering in
-                isHoveredRedirectButton = isHovering
-            }
-        }
-    }
-    
-    private var addModelSection: some View {
-        VStack(alignment: .center, spacing: 4) {
-            Text("Have your own API key or local model?")
-                .styleText(
-                    size: 13,
-                    weight: .regular,
-                    color: .gray100,
-                    align: .center
-                )
-            
-            Button {
-                appState.openModelSettingsTab(openSettings)
-            } label: {
-                Text("Add it here to access Onit.")
-                    .styleText(
-                        size: 13,
-                        weight: .regular,
-                        align: .center,
-                        underline: isHoveredAddModelButton
-                    )
-                    .onHover { isHovering in
-                        isHoveredAddModelButton = isHovering
-                    }
-            }
-        }
-    }
-    
-    private var emailAuthTokenForm: some View {
-        VStack(alignment: .center, spacing: 0) {
-            Image("Mail").padding(.bottom, 16)
-            
-            Text("Check your email")
-                .styleText(size: 23)
-            
-            VStack(alignment: .center, spacing: 2) {
-                Text("Click the link we sent to:")
-                    .styleText(
-                        size: 13,
-                        weight: .regular,
-                        color: .gray100
-                    )
-                
-                Text(email).styleText(size: 15, weight: .regular)
-            }
-            .padding(.vertical, 16)
-            
-            Spacer()
-            
-            VStack(alignment: .center, spacing: 28) {
-                VStack(alignment: .center, spacing: 0) {
-                    Text("Didn't get it? Check your spam folder and the spelling")
-                        .styleText(size: 12, color: .gray200, align: .center)
-                        .frame(maxWidth: .infinity)
-                    
-                    HStack(spacing: 4) {
-                        Text("of your email address, or")
-                            .styleText(size: 12, color: .gray200)
-                        
-                        Button {
-                            requestEmailLoginLink()
-                        } label: {
-                            Text("Resend Link")
-                                .styleText(size: 12, underline: isHoveredResendLinkButton)
-                        }
-                        .buttonStyle(PlainButtonStyle())
-                        .onHover { isHovering in
-                            isHoveredResendLinkButton = isHovering
-                        }
-                    }
-                }
-                
-                Text("← Back")
-                    .frame(height: 40)
-                    .padding(.horizontal, 14)
-                    .background(isHoveredBackButton ? .gray900 : .clear)
-                    .addBorder(cornerRadius: 9, stroke: .gray700)
-                    .scaleEffect(isPressedBackButton ? 0.98 : 1)
-                    .opacity(isPressedBackButton ? 0.7 : 1)
-                    .addAnimation(dependency: isHoveredBackButton)
-                    .onHover{ isHovering in isHoveredBackButton = isHovering }
-                    .gesture(
-                        DragGesture(minimumDistance: 0)
-                            .onChanged {_ in isPressedBackButton = true }
-                            .onEnded{ _ in
-                                AnalyticsManager.Auth.cancelled(provider: "email")
-                                isPressedBackButton = false
-                                requestedEmailLogin = false
-                            }
-                    )
-            }
-        }
-        .padding(.bottom, 53)
-    }
+    // MARK: - Child Components
     
     private var closeButton: some View {
         TextButton(
@@ -285,57 +79,9 @@ extension AuthFlow {
         .padding([.top, .horizontal], 40)
         .padding(.bottom, 53)
     }
-}
-
-
-// MARK: - Private Functions (UI)
-
-extension AuthFlow {
-    private func generateAgreementText() -> AttributedString {
-        var agreementText = AttributedString("By continuing with Google or email, you agree to our ")
-        agreementText.foregroundColor = .gray300
-        
-        var termsText = AttributedString("Terms of Service")
-        termsText.foregroundColor = .gray200
-        termsText.link = URL(string: "https://www.getonit.ai/")
-        agreementText.append(termsText)
-        
-        var andText = AttributedString(" and ")
-        andText.foregroundColor = .gray300
-        agreementText.append(andText)
-        
-        var privacyText = AttributedString("Privacy Policy")
-        privacyText.foregroundColor = .gray200
-        privacyText.link = URL(string: "https://www.getonit.ai/")
-        agreementText.append(privacyText)
-        
-        var periodText = AttributedString(".")
-        periodText.foregroundColor = .gray300
-        agreementText.append(periodText)
-        
-        return agreementText
-    }
     
-    @MainActor
-    private func handleLogin(loginResponse: LoginResponse) {
-        TokenManager.token = loginResponse.token
-        appState.account = loginResponse.account
-
-        if loginResponse.isNewAccount {
-            AnalyticsManager.Identity.identify(account: loginResponse.account)
-            useOpenAI = true
-            useAnthropic = true
-            useXAI = true
-            useGoogleAI = true
-            useDeepSeek = true
-            usePerplexity = true
-        }
-    }
-}
-
-// MARK: - Private Functions (email)
-
-extension AuthFlow {
+    // MARK: - Private Functions
+    
     private func validateEmail() -> Bool {
         let emailRegex = "^[A-Z0-9a-z._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$"
         let emailTest = NSPredicate(format: "SELF MATCHES %@", emailRegex)
@@ -351,102 +97,34 @@ extension AuthFlow {
         errorMessageEmail = ""
 
         if email.isEmpty {
-            let errorMsg = "Please enter your email"
+            let errorMessage = "Please enter your email"
             
-            AnalyticsManager.Auth.error(provider: provider, error: errorMsg)
-            errorMessageEmail = errorMsg
+            AnalyticsManager.Auth.error(provider: provider, error: errorMessage)
+            
+            errorMessageEmail = errorMessage
         } else if !validateEmail() {
-            let errorMsg = "Invalid email format"
+            let errorMessage = "Invalid email format"
             
-            AnalyticsManager.Auth.error(provider: provider, error: errorMsg)
-            errorMessageEmail = errorMsg
+            AnalyticsManager.Auth.error(provider: provider, error: errorMessage)
+            
+            errorMessageEmail = errorMessage
         } else {
             let client = FetchingClient()
             
             Task {
                 do {
                     AnalyticsManager.Auth.requested(provider: provider)
+                    
                     try await client.requestLoginLink(email: email)
                     requestedEmailLogin = true
                 } catch {
-                    let errorMsg = error.localizedDescription
+                    let errorMessage = error.localizedDescription
                     
-                    AnalyticsManager.Auth.failed(provider: provider, error: errorMsg)
-                    errorMessageEmail = errorMsg
+                    AnalyticsManager.Auth.failed(provider: provider, error: errorMessage)
+                    
+                    errorMessageEmail = errorMessage
                 }
             }
         }
-    }
-}
-
-// MARK: - Private Functions (Google, Apple)
-
-extension AuthFlow {
-    private func handleGoogleSignInButton() {
-        let provider = "google"
-        AnalyticsManager.Auth.pressed(provider: provider)
-        errorMessageAuth = ""
-        
-        guard let window = NSApp.keyWindow else { return }
-
-        AnalyticsManager.Auth.requested(provider: provider)
-        
-        GIDSignIn.sharedInstance.signIn(withPresenting: window) { result, error in
-            guard let result = result else {
-                if let error = error as? NSError, error.domain == "com.google.GIDSignIn", error.code == -5 {
-                    // The user canceled the sign-in flow
-                    AnalyticsManager.Auth.cancelled(provider: provider)
-                    return
-                } else if let error = error {
-                    let errorMsg = error.localizedDescription
-                    
-                    AnalyticsManager.Auth.error(provider: provider, error: errorMsg)
-                    errorMessageAuth = errorMsg
-                } else {
-                    let errorMsg = "Unknown Google sign in error"
-                    
-                    AnalyticsManager.Auth.error(provider: provider, error: errorMsg)
-                    errorMessageAuth = errorMsg
-                }
-                return
-            }
-
-            guard let idToken = result.user.idToken?.tokenString else {
-                let errorMsg = "Failed to get Google identity token"
-                
-                AnalyticsManager.Auth.error(provider: provider, error: errorMsg)
-                errorMessageAuth = errorMsg
-                return
-            }
-
-            Task {
-                do {
-                    let client = FetchingClient()
-                    let loginResponse = try await client.loginGoogle(idToken: idToken)
-                    handleLogin(loginResponse: loginResponse)
-                    AnalyticsManager.Auth.success(provider: provider)
-                } catch {
-                    let errorMsg = error.localizedDescription
-                    
-                    AnalyticsManager.Auth.failed(provider: provider, error: errorMsg)
-                    errorMessageAuth = errorMsg
-                }
-            }
-        }
-    }
-    
-    private func handleAppleCredential(_ authResults: ASAuthorization) async throws {
-        guard
-            let credentials = authResults.credential as? ASAuthorizationAppleIDCredential,
-            let identityToken = credentials.identityToken,
-            let identityTokenString = String(data: identityToken, encoding: .utf8)
-        else {
-            errorMessageAuth = "Failed to get Apple identity token"
-            return
-        }
-        
-        let client = FetchingClient()
-        let loginResponse = try await client.loginApple(idToken: identityTokenString)
-        handleLogin(loginResponse: loginResponse)
     }
 }
